Moved every command in its on folder.
